update build output
